refactoring refs: https://github.com/metasfresh/metasfresh/issues/18349
