+++ backend/de.metas.business/src/main/java/de/metas/uom/UpdateUOMConversionRequest.java
@@ -46,7 +46,4 @@ public class UpdateUOMConversionRequest
 	BigDecimal fromToMultiplier;
 
 	boolean catchUOMForProduct;
-
-	@Builder.Default
-	boolean active = true;
 }
