+++ backend/de.metas.business/src/main/sql/postgresql/system/43-de.metas.product/5727690_sys_gh18349_c_uom_conversion_product.sql
@@ -1,0 +1,5 @@
+DROP INDEX IF EXISTS c_uom_conversion_product
+;
+
+CREATE UNIQUE INDEX c_uom_conversion_product ON c_uom_conversion (c_uom_id, c_uom_to_id, m_product_id) WHERE IsActive = 'Y'
+;
